% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{align2region}
\alias{align2region}
\title{Newer version of alignemnt object to count over regions}
\usage{
align2region(galign, regions, chrlen_df)
}
\description{
Newer version of alignemnt object to count over regions
}
