% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bam2bin}
\alias{bam2bin}
\title{Convert bam files to bin counts}
\usage{
bam2bin(bam.file, chrlen.file, chr.count, bin.width, paired = F, ...)
}
\arguments{
\item{bam.file}{The bam file to be converted.}
}
\description{
Convert bam files to bin counts
}

