% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bin2bp}
\alias{bin2bp}
\title{Convert bin indices to base pair indices}
\usage{
bin2bp(bin, bin.width = 50)
}
\arguments{
\item{bin}{Bin indices}

\item{bin.width}{Bin width}
}
\description{
Convert bin indices to base pair indices
}

