% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bp2bin}
\alias{bp2bin}
\title{Convert base pair indices to bin indices}
\usage{
bp2bin(bp, bin.width = 50)
}
\arguments{
\item{bin.width}{Bin width}

\item{bin}{Base pair indices}
}
\description{
Convert base pair indices to bin indices
}

