% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{chr2gw}
\alias{chr2gw}
\title{Convert chromosome bin indices to genome wide bin indices.}
\usage{
chr2gw(chr, bins)
}
\arguments{
\item{chr}{vector of chr (Rle)}

\item{bins}{vector of bins}
}
\description{
Convert chromosome bin indices to genome wide bin indices.
}

