% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{chr2num}
\alias{chr2num}
\title{Convert seqnames to numeric chromosome indices. Old. For the new one, see seq2num.}
\usage{
chr2num(c)
}
\description{
Convert seqnames to numeric chromosome indices. Old. For the new one, see seq2num.
}

