% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{countReads}
\alias{countReads}
\title{Count reads in bins from GeonomicAlignments object.}
\usage{
countReads(align, chrlen.file, chr.count, bin.width, paired = F, counts = NULL)
}
\arguments{
\item{align}{A \code{\link{GenomicAlignments}} object.}

\item{paried}{A logical value indicating if alignment object is paired.}
}
\description{
Count reads in bins from GeonomicAlignments object.
}
