% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{df2bin}
\alias{df2bin}
\title{Convert reads saved as list of data.frames to counts}
\usage{
df2bin(df, chrlen_df, bin_width, paired = F, counts = NULL)
}
\description{
Convert reads saved as list of data.frames to counts
}
