% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{empirical.pvalue}
\alias{empirical.pvalue}
\title{Calculate empirical p-values}
\usage{
empirical.pvalue(null.vec, obs.vec, alternative = c("less", "greater",
  "two.sided"))
}
\arguments{
\item{null.vec}{Vector of empirical null distribution}

\item{obs.vec}{Vector of observed test statistics}

\item{alternative}{Type of alternative hypothesis}
}
\description{
Calculate empirical p-values
}
