% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gw2chr}
\alias{gw2chr}
\title{Convert vector of genome wide bin indices to a list of indices per chromosome.}
\usage{
gw2chr(bins, bin.width = 200)
}
\arguments{
\item{bins}{Vector of bin indices}

\item{bin.width}{Bin width}
}
\description{
Convert vector of genome wide bin indices to a list of indices per chromosome.
}
