% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{image.n}
\alias{image.n}
\title{Wrapper of \link{image} with additional features}
\usage{
\method{image}{n}(z, zlim, col = colorRampPalette(brewer.pal(9,
  "Blues"))(1000), na.color = grey.colors(1, 0.95), row.side = NULL,
  row.side.col = brewer.pal(8, "Dark2"), row.lab = NULL,
  row.lab.col = NULL, row.clust = F, col.clust = F,
  outside.below.color = "black", outside.above.color = "white",
  rowsep = NULL, colsep = NULL, sepcolor = grey.colors(1, 0.95),
  sepwidth = 0.02, cellnote = F, digit.format = "\%0.2f",
  text.col = "black", ...)
}
\description{
Wrapper of \link{image} with additional features
}

