% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{list2gw}
\alias{list2gw}
\title{Convert a list of indices per chromosome (must have 23 items in list) to a vector of genome wide bin indices}
\usage{
list2gw(bins.ls, bin.width = 200)
}
\arguments{
\item{bin.width}{Bin width}

\item{bins}{Vector of bin indices}
}
\description{
Convert a list of indices per chromosome (must have 23 items in list) to a vector of genome wide bin indices
}
