% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make.bins.gr}
\alias{make.bins.gr}
\title{Create a \code{\link{GRanges}} object for bins.}
\usage{
make.bins.gr(chrlen.file, bin.width)
}
\arguments{
\item{chrlen.file}{Text file for chromosome length}

\item{bin.width}{Bin width}
}
\description{
Create a \code{\link{GRanges}} object for bins.
}

