% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_gr_bins}
\alias{make_gr_bins}
\title{Create a \code{\link{GRanges}} object for bins.}
\usage{
make_gr_bins(chrlen_df, bin_width)
}
\arguments{
\item{chrlen_df}{data.frame for chromosome sizes}

\item{bin_width}{Bin width}
}
\description{
Create a \code{\link{GRanges}} object for bins.
}
