% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{seq2num}
\alias{seq2num}
\title{Convert seqnames to numeric chromosome indices. chrX mapped to 23; chrM and chrY mapped to NAs.}
\usage{
seq2num(seq, chrlen_df)
}
\description{
Convert seqnames to numeric chromosome indices. chrX mapped to 23; chrM and chrY mapped to NAs.
}
