% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ta2bin}
\alias{ta2bin}
\title{Convert tagAlign files to bin counts}
\usage{
ta2bin(ta.file, ...)
}
\arguments{
\item{ta.file}{bam The tagAlign file to be converted.}
}
\description{
Convert tagAlign files to bin counts
}
