% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{countReads}
\alias{countReads}
\title{Count reads in bins from GeonomicAlignments object.}
\usage{
countReads(align, chrlen.file, bin.width, counts = NULL)
}
\arguments{
\item{align}{GenomicAlignments (link) Object}
}
\description{
Count reads in bins from GeonomicAlignments object.
}

