% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{load.chrlen}
\alias{load.chrlen}
\title{Load chromosome length and bin counts into global environment}
\usage{
load.chrlen(chrlen_vec, bin.width)
}
\arguments{
\item{bin.width}{The bin width.}

\item{chrlen.file}{Text file for chromosome length}
}
\description{
Load chromosome length and bin counts into global environment
}
